+++ CHANGES.md
@@ -5,6 +5,9 @@
 * load in `__construct()`
 * update error email message
 * don't check our own plugin, cause it will always have a PHP fatal for redeclaring the class
+* refactor for restarting plugin update process after a fatal
+* add function for sending successful update email
+* refactor for inclusion into core
 
 #### 0.6.0 / 2022-08-05
 * update documentation